+++ core_backend/app/question_answer/speech_components/utils.py
@@ -4,7 +4,7 @@ from io import BytesIO
 from pydub import AudioSegment
 
 from ...llm_call.llm_prompts import IdentifiedLanguage
-from ...utils import get_http_client, setup_logger
+from ...utils import get_file_extension_from_mime_type, get_http_client, setup_logger
 
 logger = setup_logger("Voice utils")
 
@@ -95,3 +95,47 @@ async def post_to_internal_tts(
             audio_content = await response.read()
 
             return BytesIO(audio_content)
+
+
+async def download_file_from_url(file_url: str) -> tuple[BytesIO, str, str]:
+    """
+    Asynchronously download a file from a given URL using the
+    global aiohttp ClientSession and return its content as a BytesIO object,
+    along with its content type and file extension.
+    """
+    async with get_http_client() as client:
+        try:
+            async with client.get(file_url) as response:
+                if response.status != 200:
+                    error_content = await response.text()
+                    logger.error(f"Failed to download file: {error_content}")
+                    raise ValueError(f"Failed to download file: {error_content}")
+
+                content_type = response.headers.get("Content-Type")
+                if not content_type:
+                    logger.error("Content-Type header missing in response")
+                    raise ValueError("Unable to determine file content type")
+
+                file_stream = BytesIO(await response.read())
+                file_extension = get_file_extension_from_mime_type(content_type)
+
+        except Exception as e:
+            logger.error(f"Error during file download: {str(e)}")
+            raise ValueError(f"Unable to fetch file: {str(e)}") from None
+
+    return file_stream, content_type, file_extension
+
+
+async def post_to_speech_stt(file_path: str, endpoint_url: str) -> dict:
+    """
+    Post request the file to the speech endpoint to get the transcription
+    """
+    async with get_http_client() as client:
+        async with client.post(
+            endpoint_url, json={"stt_file_path": file_path}
+        ) as response:
+            if response.status != 200:
+                error_content = await response.json()
+                logger.error(f"Error from CUSTOM_STT_ENDPOINT: {error_content}")
+                raise ValueError(f"Error from CUSTOM_STT_ENDPOINT: {error_content}")
+            return await response.json()
